requirements.txt
faiss-cpu==1.7.4
markdown==3.5.2
langchain==0.1.1
langchain-community==0.0.13
langchain-core==0.1.12
langchain-openai==0.0.2.post1
langdetect==1.0.9
langsmith==0.0.83
python-dotenv==1.0.0
tiktoken==0.5.2
pydantic==2.5.3
fastapi==0.104.1
openai==1.28.1
starlette==0.27.0
uvicorn==0.24.0.post1
gunicorn==21.2.0